Changed new sensors to be binary_sensors.
