+++ custom_components/envisalink_new/binary_sensor.py
@@ -3,7 +3,10 @@ from __future__ import annotations
 
 import datetime
 
-from homeassistant.components.binary_sensor import BinarySensorEntity
+from homeassistant.components.binary_sensor import (
+    BinarySensorDeviceClass,
+    BinarySensorEntity,
+)
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import ATTR_LAST_TRIP_TIME
 from homeassistant.core import HomeAssistant
@@ -12,6 +15,9 @@ from homeassistant.helpers.restore_state import RestoreEntity
 from homeassistant.util import dt as dt_util
 
 from .const import (
+    CONF_PARTITION_SET,
+    CONF_PARTITIONNAME,
+    CONF_PARTITIONS,
     CONF_ZONE_SET,
     CONF_ZONENAME,
     CONF_ZONES,
@@ -22,7 +28,61 @@ from .const import (
 )
 from .helpers import find_yaml_info, generate_entity_setup_info, parse_range_string
 from .models import EnvisalinkDevice
-from .pyenvisalink.const import STATE_CHANGE_ZONE
+from .pyenvisalink.const import (
+    PANEL_TYPE_DSC,
+    PANEL_TYPE_HONEYWELL,
+    STATE_CHANGE_PARTITION,
+    STATE_CHANGE_ZONE,
+)
+
+_attribute_sensor_info = {
+    "partition": {
+        "ac_present": {
+            "name": "AC Power",
+            "panels": [PANEL_TYPE_DSC, PANEL_TYPE_HONEYWELL],
+            "icon": "mdi:power-plug",
+            "device_class": BinarySensorDeviceClass.POWER,
+        },
+        "ready": {
+            "name": "Ready",
+            "panels": [PANEL_TYPE_DSC, PANEL_TYPE_HONEYWELL],
+            "icon": "mdi:check",
+            "device_class": None,
+        },
+        "bat_trouble": {
+            "name": "Panel Battery",
+            "panels": [PANEL_TYPE_DSC, PANEL_TYPE_HONEYWELL],
+            "icon": "mdi:battery-alert",
+            "device_class": BinarySensorDeviceClass.PROBLEM,
+        },
+        "trouble": {
+            "name": "Panel Health",
+            "panels": [PANEL_TYPE_DSC, PANEL_TYPE_HONEYWELL],
+            "icon": "mdi:alert",
+            "device_class": BinarySensorDeviceClass.PROBLEM,
+        },
+        "fire": {
+            "name": "Fire",
+            "panels": [PANEL_TYPE_DSC, PANEL_TYPE_HONEYWELL],
+            "icon": "mdi:fire",
+            "device_class": BinarySensorDeviceClass.SMOKE,
+        },
+        "alarm": {
+            "name": "Alarm",
+            "panels": [PANEL_TYPE_DSC, PANEL_TYPE_HONEYWELL],
+            "icon": "mdi:alarm-light",
+            "device_class": None,
+        },
+    },
+    "zone": {
+        "low_battery": {
+            "name": "Wireless Sensor Battery",
+            "panels": [PANEL_TYPE_DSC],
+            "icon": "mdi:battery-alert",
+            "device_class": BinarySensorDeviceClass.BATTERY,
+        },
+    },
+}
 
 
 async def async_setup_entry(
@@ -32,14 +92,15 @@ async def async_setup_entry(
 ) -> None:
     """Set up the zone binary sensors based on a config entry."""
     controller = hass.data[DOMAIN][entry.entry_id]
+    entities = []
 
+    # Setup zone sensors
     zone_spec: str = entry.data.get(CONF_ZONE_SET, "")
     zone_set = parse_range_string(
         zone_spec, min_val=1, max_val=controller.controller.max_zones
     )
 
     zone_info = entry.data.get(CONF_ZONES)
-    entities = []
     if zone_set is not None:
         for zone_num in zone_set:
             zone_entry = find_yaml_info(zone_num, zone_info)
@@ -52,7 +113,42 @@ async def async_setup_entry(
             )
             entities.append(entity)
 
-        async_add_entities(entities)
+            for attr, info in _attribute_sensor_info["zone"].items():
+                if controller.controller.panel_type in info["panels"]:
+                    entity = EnvisalinkAttributeBinarySensor(
+                        hass,
+                        "zone",
+                        attr,
+                        zone_num,
+                        zone_entry,
+                        controller,
+                    )
+                    entities.append(entity)
+
+    # Setup partition sensors
+    partition_spec: str = entry.data.get(CONF_PARTITION_SET, "")
+    partition_set = parse_range_string(
+        partition_spec, min_val=1, max_val=controller.controller.max_partitions
+    )
+    partition_info = entry.data.get(CONF_PARTITIONS)
+    if partition_set is not None:
+        for part_num in partition_set:
+            part_entry = find_yaml_info(part_num, partition_info)
+
+            # Create sensors to reflect attributes tracked by pyenvisalink
+            for attr, info in _attribute_sensor_info["partition"].items():
+                if controller.controller.panel_type in info["panels"]:
+                    entity = EnvisalinkAttributeBinarySensor(
+                        hass,
+                        "partition",
+                        attr,
+                        part_num,
+                        part_entry,
+                        controller,
+                    )
+                    entities.append(entity)
+
+    async_add_entities(entities)
 
 
 class EnvisalinkBinarySensor(EnvisalinkDevice, BinarySensorEntity, RestoreEntity):
@@ -125,3 +221,51 @@ class EnvisalinkBinarySensor(EnvisalinkDevice, BinarySensorEntity, RestoreEntity
     def device_class(self):
         """Return the class of this sensor, from DEVICE_CLASSES."""
         return self._zone_type
+
+
+class EnvisalinkAttributeBinarySensor(EnvisalinkDevice, BinarySensorEntity):
+    """Representation of an Envisalink binary sensor."""
+
+    def __init__(self, hass, attr_type, evl_attr_name, index, extra_yaml_conf, controller):
+        """Initialize the sensor."""
+
+        sensor_info = _attribute_sensor_info[attr_type]
+
+        self._icon = sensor_info[evl_attr_name]["icon"]
+        self._attr_device_class = sensor_info[evl_attr_name]["device_class"]
+        self._evl_attr_type = attr_type
+        self._evl_attr_name = evl_attr_name
+        self._index = index
+
+        setup_info = generate_entity_setup_info(
+            controller,
+            attr_type,
+            index,
+            sensor_info[evl_attr_name]["name"],
+            extra_yaml_conf,
+        )
+        name = setup_info["name"]
+        self._attr_unique_id = setup_info["unique_id"]
+        self._attr_has_entity_name = setup_info["has_entity_name"]
+
+        LOGGER.debug("Setting up binary_sensor: %s", name)
+        super().__init__(
+            name,
+            controller,
+            STATE_CHANGE_PARTITION if attr_type == "partition" else STATE_CHANGE_ZONE,
+            index,
+        )
+
+    @property
+    def _info(self):
+        return self._controller.controller.alarm_state[self._evl_attr_type][self._index]
+
+    @property
+    def icon(self):
+        """Return the icon if any."""
+        return self._icon
+
+    @property
+    def is_on(self):
+        """Return true if sensor is on."""
+        return self._info["status"].get(self._evl_attr_name)
